test: close coverage gaps for rc.3 changes
- secrets_usage cluster.toml-ref join extracted to merge_cluster_refs
  and unit-tested (dedup + scoped-prefix keying) — the join behavior
  the TUI depends on was previously untested (only the capture was)
- restart_policy propagation through service_config_to_spec asserted
- generate_password length/charset sanity (CSPRNG must not be constant)
- service.rs tests moved to service_tests.rs (500-line file gate)
